==> 04_Modeling.ipynb <==
<center><span style="color:#b30000;font-size:35px;"><strong>Modeling Phase</strong></span></center>

<span style="color:#2929a3;font-size:20px;">Import Libraries</span>

import pandas as pd
import numpy as np
import plotly.express as px
from sklearn.preprocessing import StandardScaler, MinMaxScaler, RobustScaler, MultiLabelBinarizer, LabelEncoder, FunctionTransformer
from category_encoders import BinaryEncoder
from sklearn.compose import ColumnTransformer
from imblearn.pipeline import Pipeline as imb_Pipeline
from sklearn.pipeline import Pipeline
from imblearn.under_sampling import TomekLinks, RandomUnderSampler
from imblearn.combine import SMOTETomek
from imblearn.over_sampling import SMOTE
from sklearn.model_selection import cross_validate , StratifiedKFold
from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import MultinomialNB, GaussianNB, BernoulliNB
from sklearn.svm import SVC
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, VotingClassifier
from xgboost import XGBClassifier
from sklearn.decomposition import PCA
from sklearn.feature_extraction.text import TfidfVectorizer

<span style="color:#2929a3;font-size:20px;">Constants</span>

TECH_COLS = ['LanguageHaveWorkedWith',
            'DatabaseHaveWorkedWith',
            'PlatformHaveWorkedWith',
            'WebframeHaveWorkedWith',
            'MiscTechHaveWorkedWith',
            'ToolsTechHaveWorkedWith',
            'NEWCollabToolsHaveWorkedWith']

<span style="color:#2929a3;font-size:20px;">Read Dataset </span>

df = pd.read_pickle('Data/Cleaned_original_data.pkl')

df.head()

### Note
- In this dataset, I retained the null values in their original state due to the logic of the data, which allows for the presence of null values. It is possible that the user did not work with any type of databases or did not have any miscellaneous technologies .. etc

<span style="color:#2929a3;font-size:20px;">Shape of Dataset </span>

df.shape

<span style="color:#2929a3;font-size:20px;">Check Target Value Counts</span>

df['DevType'].value_counts()

<span style="color:#2929a3;font-size:20px;">Encode Target</span>

target_encoder = LabelEncoder()
df['DevType'] = target_encoder.fit_transform(df['DevType'])

<span style="color:#2929a3;font-size:20px;">Encode Features</span>

Objects_dict = {}
def Create_Encoded_Features():
    global Objects_dict
    encoded_dfs = {}
    TF_languages = TfidfVectorizer(stop_words='english')
    df_languages = pd.DataFrame(TF_languages.fit_transform(df['LanguageHaveWorkedWith'].fillna('NaN')).toarray(), columns=TF_languages.get_feature_names_out())
    encoded_dfs['LanguageHaveWorkedWith'] = df_languages
    
    TF_databases = TfidfVectorizer(stop_words='english')
    df_databases = pd.DataFrame(TF_databases.fit_transform(df['DatabaseHaveWorkedWith'].fillna('NaN')).toarray(), columns=TF_databases.get_feature_names_out())
    encoded_dfs['DatabaseHaveWorkedWith'] = df_databases
    
    TF_platforms = TfidfVectorizer(stop_words='english')
    df_platforms = pd.DataFrame(TF_platforms.fit_transform(df['PlatformHaveWorkedWith'].fillna('NaN')).toarray(), columns=TF_platforms.get_feature_names_out())
    encoded_dfs['PlatformHaveWorkedWith'] = df_platforms
    
    TF_webframes = TfidfVectorizer(stop_words='english')
    df_webframes = pd.DataFrame(TF_webframes.fit_transform(df['WebframeHaveWorkedWith'].fillna('NaN')).toarray(), columns=TF_webframes.get_feature_names_out())
    encoded_dfs['WebframeHaveWorkedWith'] = df_webframes
    
    TF_MiscTech = TfidfVectorizer(stop_words='english')
    df_MiscTech = pd.DataFrame(TF_MiscTech.fit_transform(df['MiscTechHaveWorkedWith'].fillna('NaN')).toarray(), columns=TF_MiscTech.get_feature_names_out())
    encoded_dfs['MiscTechHaveWorkedWith'] = df_MiscTech
    
    TF_tools = TfidfVectorizer(stop_words='english')
    df_tools = pd.DataFrame(TF_tools.fit_transform(df['ToolsTechHaveWorkedWith'].fillna('NaN')).toarray(), columns=TF_tools.get_feature_names_out())
    encoded_dfs['ToolsTechHaveWorkedWith'] = df_tools
    
    TF_NEWCollabTools = TfidfVectorizer(stop_words='english')
    df_NEWCollabTools = pd.DataFrame(TF_NEWCollabTools.fit_transform(df['NEWCollabToolsHaveWorkedWith'].fillna('NaN')).toarray(), columns=TF_NEWCollabTools.get_feature_names_out())
    encoded_dfs['NEWCollabToolsHaveWorkedWith'] = df_NEWCollabTools
    
    res = pd.concat(encoded_dfs, axis=1)
    Objects_dict['TF_languages'] = TF_languages
    Objects_dict['TF_databases'] = TF_databases
    Objects_dict['TF_platforms'] = TF_platforms
    Objects_dict['TF_webframes'] = TF_webframes
    Objects_dict['TF_MiscTech'] = TF_MiscTech
    Objects_dict['TF_tools'] = TF_tools
    Objects_dict['TF_NEWCollabTools'] = TF_NEWCollabTools
    return res

encoded_df = Create_Encoded_Features()

encoded_df['DevType'] = df['DevType']

encoded_df

<span style="color:#2929a3;font-size:20px;">Split Data to X and Y</span>

x = encoded_df.drop('DevType', axis=1)
y = encoded_df['DevType']

<span style="color:#2929a3;font-size:20px;">Model Selection Process</span>

Models = []
Models.append(('Logistic Regression',LogisticRegression(max_iter=10000)))
Models.append(('Knn',KNeighborsClassifier()))
Models.append(('MultinomialNB',MultinomialNB()))
Models.append(('GaussianNB',GaussianNB()))
Models.append(('BernoulliNB',BernoulliNB()))
Models.append(('Decision Tree', DecisionTreeClassifier()))
Models.append(('Random Forest',RandomForestClassifier()))
Models.append(('Voting Classifier',VotingClassifier(estimators=[('Logistic Regression',LogisticRegression(max_iter=10000)), ('Knn',KNeighborsClassifier()), ('MultinomialNB',MultinomialNB()), ('GaussianNB',GaussianNB()), ('Decision Tree', DecisionTreeClassifier()), ('Random Forest',RandomForestClassifier())])))
Models.append(('XGBoost',XGBClassifier()))

y.value_counts()

for model in Models:
    steps = []
    steps.append(('Scaler', MinMaxScaler()))
    steps.append(('UnderSampling',TomekLinks()))
    steps.append(('UnderSampling2',RandomUnderSampler(sampling_strategy={11:2000, 8:1500})))
    steps.append(('OverSampling', SMOTETomek(sampling_strategy={0:1000, 1:1000, 5:1000, 10:1000, 6:1000, 7:1000, 2:1000, 3:1000})))
    #steps.append(('PCA',PCA(n_components=0.90)))
    steps.append(model)
    pipeline = imb_Pipeline(steps=steps)
    res = cross_validate(pipeline , x, y, scoring='accuracy', cv=5, return_train_score=True, n_jobs=-1)
    print(f"Train Accuracy of {model[0]} is ",res['train_score'].mean())
    print(f"Test Accuracy of {model[0]} is ",res['test_score'].mean())
    print("*" * 70)

<span style="color:#2929a3;font-size:20px;">Create RandomForest Pipeline</span>

steps = []
steps.append(('Scaler', MinMaxScaler()))
steps.append(('UnderSampling',TomekLinks()))
steps.append(('UnderSampling2',RandomUnderSampler(sampling_strategy={11:2000, 8:1500})))
steps.append(('OverSampling', SMOTETomek(sampling_strategy={0:1000, 1:1000, 5:1000, 10:1000, 6:1000, 7:1000, 2:1000, 3:1000})))
steps.append(('Model',RandomForestClassifier()))
pipeline = imb_Pipeline(steps=steps)
pipeline.fit(x,y)

<span style="color:#2929a3;font-size:20px;">Hyperparameter Tuning</span>

from sklearn.model_selection import GridSearchCV

param = [
    {'Model__criterion':['gini','entropy'],
     'Model__max_depth':[25,30,35],
     'Model__min_samples_split':[3,4,5],
     'Model__n_estimators':[150,160,170]}
]

grid = GridSearchCV(estimator=pipeline, param_grid=param, cv=5, scoring='accuracy' , return_train_score=True, n_jobs=-1)

grid.fit(x,y)

grid.best_params_

grid.best_score_

<span style="color:#2929a3;font-size:20px;">Create Final Pipeline</span>

steps = []
steps.append(('Scaler', MinMaxScaler()))
steps.append(('UnderSampling',TomekLinks()))
steps.append(('UnderSampling2',RandomUnderSampler(sampling_strategy={11:2000, 8:1500})))
steps.append(('OverSampling', SMOTETomek(sampling_strategy={0:1000, 1:1000, 5:1000, 10:1000, 6:1000, 7:1000, 2:1000, 3:1000})))
steps.append(('Model',RandomForestClassifier(criterion='entropy', max_depth=30, min_samples_split=4, n_estimators=150)))
pipeline = imb_Pipeline(steps=steps)
pipeline.fit(x,y)

<span style="color:#2929a3;font-size:20px;">Dump The Model</span>

import joblib

df_cols_names = pd.DataFrame(columns=encoded_df.columns)
joblib.dump(df_cols_names,'df_cols_names.pkl')
joblib.dump(pipeline,'Model.pkl')
joblib.dump(target_encoder,'target_encoder.pkl')
joblib.dump(Objects_dict,'Objects_dict.pkl')
